requirements.txt
django
djangorestframework
markdown
django-filter
django-cors-headers
django-extensions
psycopg2_binary
python-dotenv
djangorestframework-simplejwt
Pillow
celery==5.2.7
redis==4.5.4
django-redis
django-debug-toolbar==3.2.4